requirements.txt
aiogram==3.4.1
google-generativeai==0.7.2
reportlab==4.1.0
python-dotenv==1.0.0
